[#56] Feature/type to type transformation
* first draft of the mapping mechanism

* first draft of the mapping mechanism

* added a mechanism for validating the mapping tuple

* removed all the 'new'

* sbt and dependency updates

* put all the supporting function for the instance management service into a separate trait

* started to shape the mapping service

* tried all the test with virtuoso, there are 3 tests not passing, very probably to the equal not working with changes in the order of attribute and/or values

* fixe3d the non passing tests with virtuoso

* added additional test, it needs to be refactored since virtuoso doesn't support rdf star

* small refactoring

* modeled the mapping relationship using the Singleton Property pattern

* modeled the mapping relationship using the Singleton Property pattern

* starting to shape the mapping application

* introduced better cache management

* dependency updates

* small correction

* small correction

* added mechanism for creating and updating instance driven by the mappings

* first draft of the mapping mechanism is done

* better checking if the type involved in the mapping are related to the MappingSoure and MappingTarget traits

* Detection for mapping cycles and checks on the update

---------

+++ domain/src/main/scala/it/agilelab/dataplatformshaper/domain/service/InstanceManagementService.scala
@@ -31,14 +31,14 @@ trait InstanceManagementService[F[_]]:
       instanceTypeName: String,
       predicate: Option[SearchPredicate],
       returnEntities: Boolean,
-      limit: Option[BigInt]
+      limit: Option[Int]
   ): F[Either[ManagementServiceError, List[String | Entity]]]
 
   def list(
       instanceTypeName: String,
       predicate: String,
       returnEntities: Boolean,
-      limit: Option[BigInt]
+      limit: Option[Int]
   ): F[Either[ManagementServiceError, List[String | Entity]]]
 
   def link(
